Start moving map draw helpers out of NewProjectMap
